Moved DescriptorImageView into it's own header
